% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_dfs}
\alias{bind_dfs}
\title{Bind a list of data frames together}
\usage{
bind_dfs(l)
}
\arguments{
\item{l}{A list of data frames}
}
\description{
Bind a list of data frames together
}
\keyword{internal}
