% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-to-scale.R
\name{convert_theta}
\alias{convert_theta}
\title{convert raw score to theta and proficiency estimates}
\usage{
convert_theta(raw_theta_tbl, name, round = TRUE)
}
\arguments{
\item{raw_theta_tbl}{table with theta values, in a column called \code{theta}. Produced by \code{raw_to_scale()}}

\item{name}{The name of the given test, e.g., \code{"ELA_G8"}}

\item{round}{Implement standard rounding, rather than R's rounding.}
}
\description{
convert raw score to theta and proficiency estimates
}
