% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_anchors}
\alias{create_anchors}
\title{Function for creating an anchor file for analysis}
\usage{
create_anchors(test, itemfile)
}
\arguments{
\item{test}{One specific test from \code{dbprocess::get_items()}}

\item{itemfile}{The items table from the database, e.g.,
\code{orextdb::db_get("Items")}}
}
\description{
This function is used internally, but is exported for spot checking. It is
called by \code{\link[=rasch]{rasch()}} when \code{anchored = TRUE} to create the anchor file
during analysis.
}
\keyword{internal}
