% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_pattern_frame}
\alias{create_pattern_frame}
\title{Create a patterned data frame for all possible raw scores}
\usage{
create_pattern_frame(item_names)
}
\arguments{
\item{item_names}{The column names (items) from which to generate the
data frame. These become the column names of the patterned data frame}
}
\value{
data frame with all possible raw scores
}
\description{
Create a patterned data frame for all possible raw scores
}
\keyword{internal}
