% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimate_abilities}
\alias{estimate_abilities}
\title{High-level function to esimate/create student estimates}
\usage{
estimate_abilities(test, single_df = TRUE)
}
\arguments{
\item{test}{Either a list of test score data from which to estimate models
or a single data frame with the item-level data. Note, this can be all
tests. It does not have to be a list of only one content area, for
example.}

\item{single_df}{Should a single data frame be returned? Defaults to
\code{FALSE}, in which case a list of data frames is returned (assuming)
\code{model_ob} is a list of models. Otherwise a single data frame is
returned, which may be generally useful, but is less useful when passing
the results to other functions in the package.}
}
\description{
This function does the modeling and pulls the student estimates
for every grade/content area (unless a specific grade/content area is
supplied). Student estimates are computed with all field-test
items removed.
}
