% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-params.R
\name{estimate_theta}
\alias{estimate_theta}
\title{Estimate person abilities}
\usage{
estimate_theta(model, full_demo_data)
}
\arguments{
\item{model}{The fitted model object. Output from \code{\link[=rasch]{rasch()}}.}

\item{full_demo_data}{The full test, including the demographic data, e.g.,
the output from \code{dbprocess::get_items()}. Note that if
\code{model_ob} is a list, this argument should be a corresponding list
with the same names. If \code{model_ob} is a single model, this argument
should be the single corresponding data frame.}
}
\value{
The \code{full_demo_data} data frame, with additional \code{theta}
and \code{theta_se} columns.
}
\description{
Estimate person abilities
}
\keyword{internal}
