% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exirt-package.R
\docType{package}
\name{exirt-package}
\alias{exirt}
\alias{exirt-package}
\title{exirt: Analyze data from the Oregon Extended Assessment}
\description{
Retrieves live data from the OR Extended database (via the dbprocess package) and analyzes the data. Provides functions for both analysis of the students (estimated using only the operational items), test items (both operational and field test) and tests. Note that all operational items are anchored to their previously estimated values (on the vertical scale) during analysis. This ensures that all newly estimated values (persons and field test items) are relative to the established vertical scale. Functions to produce and plot the test information function and test characteristic curves are also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datalorax/exirt}
  \item Report bugs at \url{https://github.com/datalorax/exirt/issues}
}

}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})

}
\keyword{internal}
