% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-params.R
\name{get_item_diffs}
\alias{get_item_diffs}
\title{Get item difficulties from a single model or list of models}
\usage{
get_item_diffs(model_ob, single_df = TRUE)
}
\arguments{
\item{model_ob}{The fitted model or list of fitted models}

\item{single_df}{Should a single data frame be returned? Defaults to
\code{FALSE}, in which case a list of data frames is returned (assuming)
\code{model_ob} is a list of models. Otherwise a single data frame is
returned, which may be generally useful, but is less useful when passing
the results to other functions in the package.}
}
\description{
Get item difficulties from a single model or list of models
}
