% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-params.R
\name{get_person_estimates}
\alias{get_person_estimates}
\title{Get person abilities}
\usage{
get_person_estimates(model_ob, full_demo_data, single_df = FALSE)
}
\arguments{
\item{model_ob}{The fitted model or list of fitted models}

\item{full_demo_data}{The full test, including the demographic data, e.g.,
the output from \code{dbprocess::get_items()}. Note that if
\code{model_ob} is a list, this argument should be a corresponding list
with the same names. If \code{model_ob} is a single model, this argument
should be the single corresponding data frame.}

\item{single_df}{Should a single data frame be returned? Defaults to
\code{FALSE}, in which case a list of data frames is returned (assuming)
\code{model_ob} is a list of models. Otherwise a single data frame is
returned, which may be generally useful, but is less useful when passing
the results to other functions in the package.}
}
\value{
Person estimates with confidence standard errors, including
conversions to RIT scores and the performance level
}
\description{
Get person abilities
}
\examples{
library(exirt)
\dontrun{
math_items <-
   dbprocess::get_items(
      content = 'Math',
      grade = 11,
      db = 'ORExt1819'
      )
rasch_mod <-
   exirt::rasch(
      test = math_items,
      omit_field_test = TRUE
      )
person_estimates <-
   get_person_estimates(
      model_ob = rasch_mod,
      full_demo_data = math_items
      )
}
}
