% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{iif}
\alias{iif}
\title{Item information function}
\usage{
iif(b, theta)
}
\arguments{
\item{b}{The item difficulty estimate}

\item{theta}{The person ability estimate. This is the \code{theta} estimate
from \code{\link[=get_person_estimates]{get_person_estimates()}} (not the RIT score).}
}
\description{
Returns the individual information for a given item along the vector of
theta values supplied. These are summed to provide the test infromation
function.
}
\keyword{internal}
