% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_item}
\alias{is_item}
\title{checks if column names in a data frame is is an item}
\usage{
is_item(d)
}
\arguments{
\item{d}{test from \code{dbprocess::get_items()}}
}
\value{
a logical vector (match or not for each element of x).
}
\description{
checks if column names in a data frame is is an item
}
\keyword{internal}
