% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prob}
\alias{prob}
\title{Estimate the probability of a correct response}
\usage{
prob(b, theta)
}
\arguments{
\item{b}{The item difficulty estimate}

\item{theta}{The person ability estimate. This is the \code{theta} estimate
from \code{\link[=get_person_estimates]{get_person_estimates()}} (not the RIT score).}
}
\description{
This function works because we know we have a Rasch model, otherwise the
function would be just a bit more complicated. Provides the probability of a
correct response for a given item difficulty, \code{b}, and a given ability
estimate, \code{theta}. Note that \code{theta} is typically passed as a
vector of abilities and so a vector of probabilities is returned.
}
\keyword{internal}
