% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-to-scale.R
\name{pull_cuts}
\alias{pull_cuts}
\title{use the \code{cuts.R} file to get the cut values for proficiency thresholds}
\usage{
pull_cuts(name)
}
\arguments{
\item{name}{The name of the given test, e.g., \code{"ELA_G8"},
\code{"Math_G11"}.}
}
\description{
use the \code{cuts.R} file to get the cut values for proficiency thresholds
}
