% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-params.R
\name{pull_item_diffs}
\alias{pull_item_diffs}
\title{Pull item difficulties from a single model}
\usage{
pull_item_diffs(model)
}
\arguments{
\item{model}{The fitted model object. Output from \code{\link[=rasch]{rasch()}}.}
}
\value{
A data frame with the item ID, difficulty, and 95\% CI.
}
\description{
Pull item difficulties from a single model
}
\examples{
library(exirt)
\dontrun{
rasch_mod_ft <-
  exirt::rasch(
    test = math_items,
    omit_field_test = FALSE
  )
exirt::get_item_diffs(rasch_mod_ft)
}
}
\keyword{internal}
