% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim2.R
\name{rasch}
\alias{rasch}
\title{Function to estimate a Rasch model}
\usage{
rasch(test, items, anchored = TRUE, ...)
}
\arguments{
\item{test}{One specific test from \code{dbprocess::get_items()}}

\item{items}{The items table from the database, e.g.,
\code{orextdb::db_get("Items")}}

\item{anchored}{Should the model be estimated with the operational items
anchored at their estimated values? Logical, defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link[TAM:tam.mml]{TAM::tam.mml()}}}
}
\description{
Uses \code{\link[TAM:tam.mml]{TAM::tam.mml()}} under the hood, becuase the theta estimates from this
package most closely matched those from Winsteps (specifically for extreme)
cases.
}
