% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{rasch_}
\alias{rasch_}
\title{Internal function to estimate a Rasch model}
\usage{
rasch_(test, items, omit_field_test = TRUE, ...)
}
\arguments{
\item{test}{One specific test from \code{dbprocess::get_items()}}

\item{items}{The items table from the database, e.g.,
\code{orextdb::db_get("Items")}}

\item{...}{Additional arguments passed to \code{\link[TAM:tam.mml]{TAM::tam.mml()}}}
}
\description{
Internal function to estimate a Rasch model
}
\keyword{internal}
