% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-to-scale.R
\name{raw_to_scale_subscores}
\alias{raw_to_scale_subscores}
\title{obtain raw to scale scores for the ELA subgroups}
\usage{
raw_to_scale_subscores()
}
\description{
obtain raw to scale scores for the ELA subgroups
}
\examples{
library(exirt)
\dontrun{
raw_to_scale_subscores()
}
}
