% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round2}
\alias{round2}
\title{Implement standard rounding, rather than R's rounding}
\usage{
round2(x, n = 0)
}
\arguments{
\item{x}{A number (type == double)}

\item{n}{The number of digits to round to}
}
\description{
Implement standard rounding, rather than R's rounding
}
\keyword{internal}
