% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tcc}
\alias{tcc}
\title{Estimate the expected total score for any values of theta}
\usage{
tcc(item_diff_table, theta = seq(-6, 6, 0.01))
}
\arguments{
\item{item_diff_table}{The data frame returned from \code{\link[=get_item_diffs]{get_item_diffs()}}
with \code{single_df = FALSE}.}

\item{theta}{The person ability estimate. This is the \code{theta} estimate
from \code{\link[=get_person_estimates]{get_person_estimates()}} (not the RIT score). Defaults to a sequence
from -6 to 6 in increments of 0.01.}
}
\description{
Internally, the theta scores are transformed to RIT scores so the resulting
output is more directly interpretable.
}
