% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tcc_}
\alias{tcc_}
\title{Create a dataframe used to make test characteristic curves}
\usage{
tcc_(item_diff_table, name, theta)
}
\arguments{
\item{item_diff_table}{The data frame returned from \code{\link[=get_item_diffs]{get_item_diffs()}}.}

\item{name}{The name of the given test, e.g., \code{"ELA_G8"},
\code{"Math_G11"}.}

\item{theta}{The person ability estimate. This is the \code{theta} estimate}
}
\description{
Create a dataframe used to make test characteristic curves
}
\keyword{internal}
