% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tcc_plot}
\alias{tcc_plot}
\title{Produce Test Characteristic Curve Plots}
\usage{
tcc_plot(
  item_diff_table,
  theta = seq(-6, 6, 0.01),
  content = c("ELA|Math", "ELA", "Math", "Science", "ELA|Science", "Math|Science",
    "all"),
  grades = "[3-8]"
)
}
\arguments{
\item{item_diff_table}{The data frame returned from \code{\link[=get_item_diffs]{get_item_diffs()}}.}

\item{theta}{The person ability estimate. This is the \code{theta} estimate
from \code{\link[=get_person_estimates]{get_person_estimates()}} (not the RIT score). Defaults to a sequence
from -6 to 6 in increments of 0.01.}

\item{content}{Subset the plot to a specific content area. Defaults to
"ELA|Math", which will select both ELA and Math.}

\item{grades}{The grade levels to include in the plot. Defaults to
\code{[3-8]}, in which case grades 3 through 8 will be selected. Any grade
range can be supplied, with the range passed in brackets. A single grade
can also be supplied, which does not require brackets, (e.g., \code{5}
would select only grade 5)}
}
\description{
Produce Test Characteristic Curve Plots
}
