% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tif_}
\alias{tif_}
\title{Test information function}
\usage{
tif_(id, name, theta = seq(-6, 6, 0.01))
}
\arguments{
\item{id}{The data frame returned from \code{\link[=get_item_diffs]{get_item_diffs()}}.}

\item{name}{The name of the given test, e.g., \code{"ELA_G8"},
\code{"Math_G11"}.}

\item{theta}{The person ability estimate. This is the \code{theta} estimate
from \code{\link[=get_person_estimates]{get_person_estimates()}} (not the RIT score). Defaults to a sequence
from -6 to 6 in increments of 0.01.}
}
\description{
Sums the item information for each \code{theta} value supplied to provide
the overall information provided by the test for the given \code{theta}
values supplied. Internally, the \code{theta} values are converted to RIT
scores to make the output more directly interpretable.
}
\keyword{Internal}
