% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tif_plot}
\alias{tif_plot}
\title{Plot the test infromation function}
\usage{
tif_plot(id, theta = seq(-6, 6, 0.01))
}
\arguments{
\item{id}{The data frame returned from \code{\link[=get_item_diffs]{get_item_diffs()}}.}

\item{theta}{The person ability estimate. This is the \code{theta} estimate
from \code{\link[=get_person_estimates]{get_person_estimates()}} (not the RIT score). Defaults to a sequence
from -6 to 6 in increments of 0.01.}
}
\description{
Plot the test infromation function
}
