% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tif_plot_}
\alias{tif_plot_}
\title{Internal function that uses \code{tif()} output}
\usage{
tif_plot_(tif_df)
}
\arguments{
\item{tif_df}{The test information function data frame. Output from \code{\link[=tif]{tif()}}}
}
\description{
Internal function that uses \code{tif()} output
}
\keyword{internal}
