% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim2.R
\name{estimate_theta}
\alias{estimate_theta}
\title{Estimate person abilities}
\usage{
estimate_theta(model, full_demo_data)
}
\arguments{
\item{model}{The fitted model. Output from \code{\link[=rasch]{rasch()}}.}

\item{full_demo_data}{The full test, including the demographic data, e.g.,
one specific test from \code{dbprocess::get_items()}.}
}
\value{
The \code{full_demo_data} data frame, with additional \code{theta}
and \code{theta_se} columns.
}
\description{
Estimate person abilities
}
