% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_test}
\alias{infer_test}
\title{Function to infer the test type based on the item ids.}
\usage{
infer_test(item_ids)
}
\arguments{
\item{item_ids}{a vector of item ids from which to infer test type}
}
\description{
Will stop and give a message if there is more than 1 content area.
}
