% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_collapse}
\alias{paste_collapse}
\title{Paste a list together}
\usage{
paste_collapse(x)
}
\arguments{
\item{x}{A character vector}
}
\description{
Adds commas after each element and a "and" between the last and second to
last elements.
}
\examples{
{
paste_collapse(c("red", "green", "blue", "orange"))
#> [1] "red, green, blue, and orange"
}
}
\keyword{internal}
