% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim2.R
\name{pull_item_diffs}
\alias{pull_item_diffs}
\title{Pull item difficulties from the model}
\usage{
pull_item_diffs(model)
}
\arguments{
\item{model}{The fitted model. Output from \code{\link[=rasch]{rasch()}}.}
}
\value{
A data frame with the item ID, difficulty, and 95\% CI.
}
\description{
Pull item difficulties from the model
}
