% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{rasch}
\alias{rasch}
\title{Function estimate Rasch models}
\usage{
rasch(test, omit_field_test = TRUE, anchored = TRUE, ...)
}
\arguments{
\item{test}{Either a list of test score data from which to estimate models
or a single data frame with the item-level data. Note, this can be all
tests. It does not have to be a list of only one content area, for
example.}

\item{omit_field_test}{Should field test items (those that are not anchored)
be omitted in the estimation? Defaults to \code{TRUE}. Note this should
always be \code{TRUE} when person estimates are extracted from the model.}

\item{anchored}{Should the model be estimated with the operational items
anchored at their estimated values? Logical, defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link[TAM:tam.mml]{TAM::tam.mml()}}}
}
\value{
A list of models
}
\description{
Uses \code{\link[TAM:tam.mml]{TAM::tam.mml()}} under the hood, becuase the theta estimates from this
package most closely matched those from Winsteps (specifically for extreme)
cases. Can estimate models from either a list tests or a single test.
Takes care of the item prep (i.e., recoding \code{NA} values to 0) and,
optionally, item anchoring.
}
\examples{
\dontrun{
  library(exirt)
  all_tests <- dbprocess::get_items(db = "2021")
  models <- rasch_batch(all_tests)
}
}
